+++ library/include/ck/library/tensor_operation_instance/gpu/grouped_convolution_backward_data.hpp
@@ -253,15 +253,6 @@ struct DeviceOperationInstanceFactory<
                         op_ptrs);
                 }
 #endif
-#if defined CK_ENABLE_FP16 && defined CK_ENABLE_FP8 && defined CK_ENABLE_BF8
-                if constexpr(is_same_v<InDataType, F16> && is_same_v<WeiDataType, F16> &&
-                             is_same_v<OutDataType, F16> && is_same_v<ComputeTypeA, bf8_t> &&
-                             is_same_v<ComputeTypeB, f8_t>)
-                {
-                    add_device_grouped_conv3d_bwd_data_xdl_ndhwgk_gkzyxc_ndhwgc_input_f16_comp_bf8f8_instances(
-                        op_ptrs);
-                }
-#endif
 #ifdef CK_ENABLE_FP32
                 if constexpr(is_same_v<InDataType, F32> && is_same_v<WeiDataType, F32> &&
                              is_same_v<OutDataType, F32> && is_same_v<ComputeTypeA, F32> &&